requirements.txt
asgiref==3.8.1
Django==5.1.7
django-phonenumber-field==8.0.0
djangorestframework==3.17.1
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.3
tzdata==2025.1
anthropic>=0.40.0
pydantic>=2.0
pytest>=8.0
pytest-mock>=3.0
django-cors-headers==4.3.1